Assert slice lengths are always > 0 (#5)
This commit is currently necessary to fix a soundness hole on Windows, and it's
also necessary to provide future compatibility for soundness holes on Unix.

This unsettling issue was first reported in #4. After some investigation I
think I understand why this is happening. Right now this crate sort of "abuses"
the compiler to get `&IoVec` and `&mut IoVec` to have the exact same in-memory
representation as the underlying networking buffers. To do this the type is
defined as `IoVec([u8])` (morally). The problem with this representation is
that we don't know where the pointer/length are in the two-word representation,
and on Unix/Windows it's different.

Now Rust in general must uphold the guarantee that the pointer part of `&[u8]`
is never null. This means that on one platform the pointer part of `&IoVec`
will actually be the data pointer, but on the other platform the pointer part
is actually the length (see the problem coming?). Turns out on Unix the current
representation of `&[u8]` aligns with the representation of `iovec`. On
Windows, however, the fields are swapped, where the pointer of `&[u8]` is the
length of `WSABUF`.

So what's happening here is that we're accidentally violating a language
guarantee. By creating a zero-length `WSABUF` we're effectively calling
`slice::from_raw_parts(ptr::null(), some_random_pointer as usize)`, and that's
undefined behavior!

This commit closes this soundness hole by adding asserts to both modules that
the slice lengths used here are always bigger than 0. This is, unfortunately,
a backwards incompatible change. It is, however, required for soundness on
Windows and future compatibility on Unix as we're not sure the slice
representation will stay the same.

Ideally this commit would also remove the `Default` implementations, but I've
opted for now to just go for behavior changes as opposed to API-breaking
changes.

Closes #4