Fix undefined behavior when transmuting slices.
The representation of slices is unspecified meaning that a mem::transmute is not
guaranteed to do the right thing although it currently does in practice so AFAICT
this currently is not a security vulnerability.

The fix is "simple": instead of transmuting the slices, we get the pointer and
the length of a slice, and use `from_raw_parts` to construct the other by
casting the pointer to the other pointer type. For this to work correctly, both
types have to have the same size, since otherwise, the length of the first slice
would be incorrect. Also, either both types have the same alignment, or the
original pointer is suitably aligned for the second type. Since the types here
are fixed and internal to the library, we assert these conditions using
`debug_assert!`s only. As long as these transmutes are tested, any changes in
layout that would introduce undefined behavior will be detected.